requirements.txt
pyyaml >= 5.3.1
transformers==4.9.2
datasets==1.11.0
torch==1.6.0
pymongo==3.12.0
pandas>=1.1.5
scikit_learn==0.24.2
Flask==2.0.1